Moved ImagePointDescriptionOutput from modeling_outputs.py to modeling_superpoint.py
